miner cleanup
